Edit form structure
