+++ src/actions/capitalDistributor/types/capitalDistributorActionCreateCampaign.ts
@@ -1,0 +1,19 @@
+import type { IProposalAction } from '@aragon/gov-ui-kit';
+import type { ICapitalDistributorCreateCampaignFormData } from '../components/capitalDistributorCreateCampaignActionCreate';
+import type { CapitalDistributorActionType } from './enum/capitalDistributorActionType';
+
+export interface ICapitalDistributorActionCreateCampaign
+    extends Omit<IProposalAction, 'type'> {
+    /**
+     * The type of the proposal action.
+     */
+    type: CapitalDistributorActionType.CREATE_CAMPAIGN;
+    /**
+     * DAO id.
+     */
+    daoId: string;
+    /**
+     * Campaign details collected from the action form in create phase.
+     */
+    campaignDetails?: ICapitalDistributorCreateCampaignFormData;
+}
